Rename: 아토믹 디자인패턴으로 폴더구조 변경, 사용안하는 import 삭제
