+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import "./globals.css";
 import { Open_Sans } from "next/font/google";
-import Header from "@/components/Header";
+import Header from "@/components/organisms/Header";
 import AuthContext from "../context/AuthContext";
 import { YoutubeApiProvider } from "@/context/YoutubeApiContext";
 import ConsertApiProvider from "@/context/ConsertApiContext";
