+++ src/app/page.tsx
@@ -1,7 +1,5 @@
-import ArtistCart from '@/components/ArtistCard'
-import HotRank from '@/components/HotRank'
-import MyChoice from '@/components/MyChoice'
-import Image from 'next/image'
+import HotRank from "@/components/organisms/HotRank";
+import MyChoice from "@/components/organisms/MyChoice";
 
 export default function Home() {
   return (
@@ -9,5 +7,5 @@ export default function Home() {
       <HotRank />
       <MyChoice />
     </section>
-  )
+  );
 }
