+++ src/components/ArtistWrap.tsx
@@ -1,18 +1,12 @@
 "use client";
 import Artist from "@/service/artist/Artist";
-import { RecentConsertData } from "@/types/_type";
+import { ArtistData } from "@/types/_type";
 import React, { useState } from "react";
 import useSWR from "swr";
-import ArtistCard from "./ArtistCard";
+import ArtistCard from "./molecules/ArtistCard";
 
 
-export type ArtistData = {
-  _id: string;
-  profile: string;
-  recentConserts: RecentConsertData[];
-  korName: string;
-  enName: string;
-};
+
 
 export default function ArtistWrap() {
   const [page, setPage] = useState(1);
