--- src/components/MyChoice.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import CarouselView from './ui/CarouselView';
-
-export default function MyChoice() {
-    return (
-        <section className='my-8'>
-            <h2 className='text-2xl font-bold'>My-Choice</h2>
-        </section>
-    );
-}
-
