+++ src/components/molecules/AlbumCard.tsx
@@ -1,7 +1,7 @@
 import { AlbumData } from "@/types/_type";
 import Image from "next/image";
 import React from "react";
-import Title from "./common/Title";
+import Title from "../atoms/Title";
 
 export default function AlbumCard({ album }: { album: AlbumData }) {
   const {
