+++ src/components/organisms/HotRank.tsx
@@ -3,10 +3,8 @@ import { useConsertApi } from "@/context/ConsertApiContext";
 import { Rank } from "@/types/_type";
 import React from "react";
 import useSWR from "swr";
-import ConsertCard from "./ConsertCard";
-import CarouselView from "./ui/CarouselView";
-
-
+import ConsertCard from "../molecules/ConsertCard";
+import CarouselView from "./CarouselView";
 
 export default function HotRank() {
   const consert = useConsertApi();
@@ -21,7 +19,7 @@ export default function HotRank() {
     <section className="mt-8">
       <h2 className="text-2xl font-bold">TOP 10</h2>
       {ranks && (
-        <CarouselView type={'auto'}>
+        <CarouselView type={"auto"}>
           {ranks.map((rank) => (
             <ConsertCard key={rank._id} consert={rank} />
           ))}
