+++ src/components/organisms/MyChoice.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import CarouselView from "./CarouselView";
+
+export default function MyChoice() {
+  return (
+    <section className="my-8">
+      <h2 className="text-2xl font-bold">My-Choice</h2>
+    </section>
+  );
+}
